requirements.txt
torch>=2.1.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
selenium>=4.15.0
beautifulsoup4>=4.12.2
aiohttp>=3.9.0
python-nmap>=0.7.1
colorama>=0.4.6
python-dotenv>=1.0.0
pillow>=10.0.0